small irc changes
This should make the irc module more to itself.
