+++ main.py
@@ -86,7 +86,7 @@ customStart = ""
 
 ####variables
 config = {"channelName": "", "pageToken": "", "serverName": "", "discordToken": "","discordToIRCFormating": "", "IRCToDiscordFormatting":""}
-from modules.irc import irc
+from modules import irc
 botName = "none"
 
 
@@ -356,38 +356,7 @@ if firstRun == "on":
         
         
 
-#this starts everything for the irc client 
-##possibly could of put all this in a class and been done with it?
-def ircStart():
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-    #loop = asyncio.get_event_loop(loop)
-    Twitch_boT = irc()
-    loop.create_task(Twitch_boT.irc_bot(loop))
-    loop.run_forever()
-    loop.close()
 
-def ircCheck():
-    global config
-    ircThread = threading.Thread(target=ircStart) #creates the thread for the irc client
-    ircThread.start() #starts the irc bot
-    time.sleep(10)
-    # while True:
-        # time.sleep(1)
-        # state = ircThread.isAlive()
-        # if state == False:
-            # print("damn it")
-            # ircThread = threading.Thread(target=ircStart) #creates the thread for the irc client
-            # ircThread.start() #starts the irc bot   
-        #irc msg handler
-        # j = 0
-        # for msg in variables.processedMSG: #this cycles through the array for messages unsent to irc and sends them
-            # #print(msg["sendTo"])
-            # if msg["sent"] == False and msg["sendTo"]["Bot"] == "IRC":
-                # ircClient.message(msg["sendTo"]["Channel"],msg["msgFormated"])#sends the message to the irc from whatever
-                # variables.processedMSG[j]["sent"] = True#promptly after sets that to the delete code
-            # j = j + 1
-        
 #youtube
         
 botName = "none"
@@ -428,7 +397,7 @@ print("test")
 chatControlThread = threading.Thread(target=mainBot.mainBot().main)
 chatControlThread.start()
 
-ircCheckThread = threading.Thread(target=ircCheck)#starts my irc check thread which should print false if the irc thread dies.
+ircCheckThread = threading.Thread(target=irc.ircCheck)#starts my irc check thread which should print false if the irc thread dies.
 if config["Bot"]["IRC"]["Enabled"] == True:
     ircCheckThread.start()
     print("IRC Loaded")
